requirements.txt
django 
djangorestframework 
mysqlclient
djangorestframework-simplejwt 
django-filter
python-dotenv
Pillow